chore: rename stockTransfers dictionary file to be camelCase
